+++ docs/components/ToolbarButton/ToolbarButtonRenderer.tsx
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+const ToolbarButtonRenderer: React.FC = () => <></>;
+
+export default ToolbarButtonRenderer;