+++ styleguide.config.js
@@ -1,5 +1,4 @@
 const path = require('path');
-const resolve = require('resolve');
 
 module.exports = {
   title: 'Synerise Design System',
@@ -10,110 +9,50 @@ module.exports = {
   sections: [
     {
       name: 'Intro',
-      content: './docs/intro.md',
+      content: 'docs/intro.md'
     },
     {
-      name: 'Pallete',
-      content: './docs/palette.md',
+      name: 'Colors' ,
+      content: 'docs/colors.md'
     },
     {
       name: 'Components',
-      components: () => ([
-        './packages/components/Button/**/*.tsx',
-        './packages/components/Avatar/**/*.tsx',
-        './packages/components/Modal/**/*.tsx',
-      ])
-    },
-    {
-      name: 'Components #2',
-      components: () => ([
-        './packages/components/AppMenu/**/*.tsx',
-        './packages/components/Badge/**/*.tsx',
-      ])
+      components: () => {
+        return [
+          './packages/components/avatar/src/Avatar.tsx',
+          './packages/components/button/src/Button.tsx'
+        ]
+      }
     }
   ],
-  webpackConfig: {
-    module: {
-      rules: [
-        {
-          test: /\.(tsx|jsx|ts|js)?$/,
-          exclude: /node_modules/,
-          use: [
-            {
-              loader: 'babel-loader',
-              options: {
-                presets: [
-                  '@babel/preset-react',
-                  [
-                    '@babel/preset-env',
-                    {
-                      modules: false,
-                    },
-                  ],
-                  '@babel/preset-typescript',
-                ],
-                plugins: [
-                  'babel-plugin-styled-components',
-                  '@babel/plugin-proposal-object-rest-spread',
-                  ['@babel/plugin-proposal-class-properties', { loose: true }],
-                  [
-                    'transform-rename-import',
-                    {
-                      replacements: [
-                        {
-                          original: '^(.+?)\\.jsx$',
-                          replacement: '$1.js',
-                        },
-                      ],
-                    },
-                  ],
-                ],
-              }
-            }
-          ]
-        },
-        {
-          test: /\.less$/,
-          use: [
-            {
-              loader: 'style-loader',
-            },
-            {
-              loader: 'css-loader',
-            },
-            {
-              loader: 'less-loader',
-              options: {
-                javascriptEnabled: true
-              }
-            }]
-        },
-        {
-          test: /\.css$/,
-          use: [
-            {
-              loader: 'style-loader',
-            },
-            {
-              loader: 'css-loader',
-            }]
-        },
-        {
-          test: /\.(woff(2)?|ttf|eot|svg|jpg|png|svg)(\?v=\d+\.\d+\.\d+)?$/,
-          use: [
-            {
-              loader: 'file-loader',
-              options: {
-                name: '[name].[ext]',
-              }
-            }
-          ]
-        }
-      ]
+  usageMode: 'expand',
+  theme: {
+    color: {
+      base: '#6a7580',
+      light: '#949ea6',
+      lightest: '#b5bdc3',
+      link: '#0b68ff',
+      linkHover: '#0044d9',
+      focus: '#0044d9',
+      border: '#dbe0e3',
+      name: '#690',
+      type: '#905',
+      error: '#c00',
+      baseBackground: '#fff',
+      codeBackground: '#f5f5f5',
+      sidebarBackground: '#f5f5f5',
+      ribbonBackground: '#e90',
+      ribbonText: '#fff',
     }
   },
-  ignore: ['**/*.styles.tsx', '**/*.spec.tsx', '**/utils/**/*', '**/*.types.tsx', '**/*.context.tsx', '**/icon/**/*', 'packages/components/**/dist/*.js'],
+  verbose: true,
+  skipComponentsWithoutExample: true,
+  webpackConfig: require('./config/styleguidist/webpack.config.js'),
+  ignore: ['**/*.styles.tsx', '**/*.spec.tsx', '**/utils/**/*', '**/node_modules/**/*', '**/*.types.tsx', '**/*.context.tsx', '**/icon/**/*', 'packages/components/**/dist/*.js'],
   styleguideComponents: {
-    Wrapper: path.join(__dirname, 'docs/components/Wrapper'),
+    Wrapper: path.join(__dirname, 'docs/components/Wrapper/Wrapper'),
+    ToolbarButtonRenderer: path.join(__dirname, 'docs/components/ToolbarButton/ToolbarButtonRenderer'),
+    ReactComponentRenderer: path.join(__dirname, 'docs/components/ReactComponent/ReactComponentRenderer'),
+    Props: path.join(__dirname, 'docs/components/Props/PropsRenderer'),
   }
 }