+++ packages/components/app-menu/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.30](https://github.com/Synerise/synerise-design/compare/@synerise/ds-app-menu@0.3.29...@synerise/ds-app-menu@0.3.30) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-app-menu
+
+
+
+
+
+## [0.3.29](https://github.com/Synerise/synerise-design/compare/@synerise/ds-app-menu@0.3.28...@synerise/ds-app-menu@0.3.29) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-app-menu
+
+
+
+
+
+## [0.3.28](https://github.com/Synerise/synerise-design/compare/@synerise/ds-app-menu@0.3.27...@synerise/ds-app-menu@0.3.28) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-app-menu
+
+
+
+
+
 ## [0.3.27](https://github.com/Synerise/synerise-design/compare/@synerise/ds-app-menu@0.3.26...@synerise/ds-app-menu@0.3.27) (2020-03-27)
 
 **Note:** Version bump only for package @synerise/ds-app-menu