+++ packages/components/autocomplete/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-autocomplete@0.3.10...@synerise/ds-autocomplete@0.3.11) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-autocomplete
+
+
+
+
+
+## [0.3.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-autocomplete@0.3.9...@synerise/ds-autocomplete@0.3.10) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-autocomplete
+
+
+
+
+
+## [0.3.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-autocomplete@0.3.8...@synerise/ds-autocomplete@0.3.9) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-autocomplete
+
+
+
+
+
 ## [0.3.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-autocomplete@0.3.7...@synerise/ds-autocomplete@0.3.8) (2020-03-27)
 
 **Note:** Version bump only for package @synerise/ds-autocomplete