+++ packages/components/avatar/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.8.2...@synerise/ds-avatar@0.8.3) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-avatar
+
+
+
+
+
+## [0.8.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.8.1...@synerise/ds-avatar@0.8.2) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-avatar
+
+
+
+
+
+## [0.8.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.8.0...@synerise/ds-avatar@0.8.1) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-avatar
+
+
+
+
+
 # [0.8.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-avatar@0.7.8...@synerise/ds-avatar@0.8.0) (2020-04-06)