+++ packages/components/button-group/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.15](https://github.com/Synerise/synerise-design/compare/@synerise/ds-button-group@0.5.14...@synerise/ds-button-group@0.5.15) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-button-group
+
+
+
+
+
+## [0.5.14](https://github.com/Synerise/synerise-design/compare/@synerise/ds-button-group@0.5.13...@synerise/ds-button-group@0.5.14) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-button-group
+
+
+
+
+
+## [0.5.13](https://github.com/Synerise/synerise-design/compare/@synerise/ds-button-group@0.5.12...@synerise/ds-button-group@0.5.13) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-button-group
+
+
+
+
+
 ## [0.5.12](https://github.com/Synerise/synerise-design/compare/@synerise/ds-button-group@0.5.11...@synerise/ds-button-group@0.5.12) (2020-04-06)
 
 **Note:** Version bump only for package @synerise/ds-button-group