+++ packages/components/button/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-button@0.11.2...@synerise/ds-button@0.11.3) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-button
+
+
+
+
+
+## [0.11.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-button@0.11.1...@synerise/ds-button@0.11.2) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-button
+
+
+
+
+
+## [0.11.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-button@0.11.0...@synerise/ds-button@0.11.1) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-button
+
+
+
+
+
 # [0.11.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-button@0.10.0...@synerise/ds-button@0.11.0) (2020-04-06)