+++ packages/components/card-tabs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.19](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.6.18...@synerise/ds-card-tabs@0.6.19) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-card-tabs
+
+
+
+
+
+## [0.6.18](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.6.17...@synerise/ds-card-tabs@0.6.18) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-card-tabs
+
+
+
+
+
+## [0.6.17](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.6.16...@synerise/ds-card-tabs@0.6.17) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-card-tabs
+
+
+
+
+
 ## [0.6.16](https://github.com/Synerise/synerise-design/compare/@synerise/ds-card-tabs@0.6.15...@synerise/ds-card-tabs@0.6.16) (2020-04-06)
 
 **Note:** Version bump only for package @synerise/ds-card-tabs