+++ packages/components/drawer/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.13](https://github.com/Synerise/synerise-design/compare/@synerise/ds-drawer@0.4.12...@synerise/ds-drawer@0.4.13) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-drawer
+
+
+
+
+
+## [0.4.12](https://github.com/Synerise/synerise-design/compare/@synerise/ds-drawer@0.4.11...@synerise/ds-drawer@0.4.12) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-drawer
+
+
+
+
+
+## [0.4.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-drawer@0.4.10...@synerise/ds-drawer@0.4.11) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-drawer
+
+
+
+
+
+## [0.4.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-drawer@0.4.9...@synerise/ds-drawer@0.4.10) (2020-04-10)
+
+
+### Bug Fixes
+
+* **drawer:** removed inline styles ([d0b9b87](https://github.com/Synerise/synerise-design/commit/d0b9b874a47e024495772741cf282fc32154decc))
+
+
+
+
+
 ## [0.4.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-drawer@0.4.8...@synerise/ds-drawer@0.4.9) (2020-04-02)