+++ packages/components/inline-edit/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.32](https://github.com/Synerise/synerise-design/compare/@synerise/ds-inline-edit@0.1.31...@synerise/ds-inline-edit@0.1.32) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-inline-edit
+
+
+
+
+
+## [0.1.31](https://github.com/Synerise/synerise-design/compare/@synerise/ds-inline-edit@0.1.30...@synerise/ds-inline-edit@0.1.31) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-inline-edit
+
+
+
+
+
+## [0.1.30](https://github.com/Synerise/synerise-design/compare/@synerise/ds-inline-edit@0.1.29...@synerise/ds-inline-edit@0.1.30) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-inline-edit
+
+
+
+
+
 ## [0.1.29](https://github.com/Synerise/synerise-design/compare/@synerise/ds-inline-edit@0.1.28...@synerise/ds-inline-edit@0.1.29) (2020-04-01)