+++ packages/components/input/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.15](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.11.14...@synerise/ds-input@0.11.15) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
+## [0.11.14](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.11.13...@synerise/ds-input@0.11.14) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
+## [0.11.13](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.11.12...@synerise/ds-input@0.11.13) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-input
+
+
+
+
+
 ## [0.11.12](https://github.com/synerise/synerise-design/compare/@synerise/ds-input@0.11.11...@synerise/ds-input@0.11.12) (2020-03-27)
 
 **Note:** Version bump only for package @synerise/ds-input