+++ packages/components/menu/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-menu@0.1.2...@synerise/ds-menu@0.1.3) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-menu
+
+
+
+
+
+## [0.1.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-menu@0.1.1...@synerise/ds-menu@0.1.2) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-menu
+
+
+
+
+
+## [0.1.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-menu@0.1.0...@synerise/ds-menu@0.1.1) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-menu
+
+
+
+
+
+# [0.1.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-menu@0.0.7...@synerise/ds-menu@0.1.0) (2020-04-10)
+
+
+### Features
+
+* **menu:** MenuItems as Menu childs components ([e703360](https://github.com/Synerise/synerise-design/commit/e7033606927f6c303bcbae835219b5528be09b05))
+
+
+
+
+
 ## [0.0.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-menu@0.0.6...@synerise/ds-menu@0.0.7) (2020-04-09)