+++ packages/components/navbar/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.12](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.4.11...@synerise/ds-navbar@0.4.12) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-navbar
+
+
+
+
+
+## [0.4.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.4.10...@synerise/ds-navbar@0.4.11) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-navbar
+
+
+
+
+
+## [0.4.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.4.9...@synerise/ds-navbar@0.4.10) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-navbar
+
+
+
+
+
 ## [0.4.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-navbar@0.4.8...@synerise/ds-navbar@0.4.9) (2020-03-27)