+++ packages/components/scrollbar/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-scrollbar@0.0.9...@synerise/ds-scrollbar@0.0.10) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-scrollbar
+
+
+
+
+
+## [0.0.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-scrollbar@0.0.8...@synerise/ds-scrollbar@0.0.9) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-scrollbar
+
+
+
+
+
+## [0.0.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-scrollbar@0.0.7...@synerise/ds-scrollbar@0.0.8) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-scrollbar
+
+
+
+
+
 ## [0.0.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-scrollbar@0.0.6...@synerise/ds-scrollbar@0.0.7) (2020-03-27)
 
 **Note:** Version bump only for package @synerise/ds-scrollbar