+++ packages/components/select/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.17](https://github.com/synerise/synerise-design/compare/@synerise/ds-select@0.10.16...@synerise/ds-select@0.10.17) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-select
+
+
+
+
+
+## [0.10.16](https://github.com/synerise/synerise-design/compare/@synerise/ds-select@0.10.15...@synerise/ds-select@0.10.16) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-select
+
+
+
+
+
+## [0.10.15](https://github.com/synerise/synerise-design/compare/@synerise/ds-select@0.10.14...@synerise/ds-select@0.10.15) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-select
+
+
+
+
+
 ## [0.10.14](https://github.com/synerise/synerise-design/compare/@synerise/ds-select@0.10.13...@synerise/ds-select@0.10.14) (2020-04-07)