+++ packages/components/slider/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.14](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.7.13...@synerise/ds-slider@0.7.14) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+
+
+
+
+## [0.7.13](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.7.12...@synerise/ds-slider@0.7.13) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+
+
+
+
+## [0.7.12](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.7.11...@synerise/ds-slider@0.7.12) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-slider
+
+
+
+
+
 ## [0.7.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-slider@0.7.10...@synerise/ds-slider@0.7.11) (2020-03-27)
 
 **Note:** Version bump only for package @synerise/ds-slider