+++ packages/components/tabs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.24](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.9.23...@synerise/ds-tabs@0.9.24) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
+
+
+
+
+## [0.9.23](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.9.22...@synerise/ds-tabs@0.9.23) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
+
+
+
+
+## [0.9.22](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.9.21...@synerise/ds-tabs@0.9.22) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-tabs
+
+
+
+
+
 ## [0.9.21](https://github.com/synerise/synerise-design/compare/@synerise/ds-tabs@0.9.20...@synerise/ds-tabs@0.9.21) (2020-04-08)
 
 **Note:** Version bump only for package @synerise/ds-tabs