+++ packages/components/utils/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.17](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.8.16...@synerise/ds-utils@0.8.17) (2020-04-15)
+
+**Note:** Version bump only for package @synerise/ds-utils
+
+
+
+
+
+## [0.8.16](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.8.15...@synerise/ds-utils@0.8.16) (2020-04-14)
+
+**Note:** Version bump only for package @synerise/ds-utils
+
+
+
+
+
+## [0.8.15](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.8.14...@synerise/ds-utils@0.8.15) (2020-04-13)
+
+**Note:** Version bump only for package @synerise/ds-utils
+
+
+
+
+
 ## [0.8.14](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.8.13...@synerise/ds-utils@0.8.14) (2020-03-27)
 
 **Note:** Version bump only for package @synerise/ds-utils